refactor: move Reciter type to shared module
