+++ lib/audio/reciters.ts
@@ -1,4 +1,4 @@
-import type { Reciter } from '@/presentation/shared/player/types';
+import type { Reciter } from '@/types';
 
 export const RECITERS: Reciter[] = [
   { id: 2, name: 'Abdul Basit (Murattal)', path: 'AbdulBaset/Murattal' },
